+++ ThreeDSecure/src/main/java/com/braintreepayments/api/ThreeDSecureResult.java
@@ -19,7 +19,7 @@ public class ThreeDSecureResult implements Parcelable {
     private static final String PAYMENT_METHOD_KEY = "paymentMethod";
     private static final String LOOKUP_KEY = "lookup";
 
-    private ThreeDSecureNonce tokenizedCard;
+    private ThreeDSecureNonce threeDSecureNonce;
     private String errorMessage;
 
     private ThreeDSecureLookup lookup;
@@ -37,7 +37,7 @@ public class ThreeDSecureResult implements Parcelable {
 
         JSONObject cardJson = json.optJSONObject(PAYMENT_METHOD_KEY);
         if (cardJson != null) {
-            result.tokenizedCard = ThreeDSecureNonce.fromJSON(cardJson);
+            result.threeDSecureNonce = ThreeDSecureNonce.fromJSON(cardJson);
         }
 
         if (json.has(ERRORS_KEY)) {
@@ -62,12 +62,12 @@ public class ThreeDSecureResult implements Parcelable {
      * @return The {@link ThreeDSecureNonce} associated with the 3D Secure authentication
      */
     @Nullable
-    public ThreeDSecureNonce getTokenizedCard() {
-        return tokenizedCard;
+    public ThreeDSecureNonce getThreeDSecureNonce() {
+        return threeDSecureNonce;
     }
 
-    void setTokenizedCard(@Nullable ThreeDSecureNonce cardNonce) {
-        tokenizedCard = cardNonce;
+    void setThreeDSecureNonce(@Nullable ThreeDSecureNonce cardNonce) {
+        threeDSecureNonce = cardNonce;
     }
 
     /**
@@ -99,13 +99,13 @@ public class ThreeDSecureResult implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeParcelable(tokenizedCard, flags);
+        dest.writeParcelable(threeDSecureNonce, flags);
         dest.writeString(errorMessage);
         dest.writeParcelable(lookup, flags);
     }
 
     private ThreeDSecureResult(Parcel in) {
-        tokenizedCard = in.readParcelable(CardNonce.class.getClassLoader());
+        threeDSecureNonce = in.readParcelable(CardNonce.class.getClassLoader());
         errorMessage = in.readString();
         lookup = in.readParcelable(ThreeDSecureLookup.class.getClassLoader());
     }